refactor: extract banner calculation logic into a composable for better reusability and organization
